Renamed nveto -> veto pulse_processing
